+++ extensions/ploi/src/Site.tsx
@@ -47,8 +47,7 @@ const SiteListItem = ({ site, server }: { site: ISite; server: IServer }) => {
         site.phpVersion ? "PHP version: " + String(site.phpVersion) ?? "" : ""
       }
       icon={stateIcon}
-      accessoryIcon={siteStatusState(site).icon}
-      accessoryTitle={stateText}
+      accessories={[{icon: siteStatusState(site).icon}, {text: stateText}]}
       actions={
         <ActionPanel>
           <ActionPanel.Section>
@@ -94,7 +93,7 @@ export const SitesSingleView = ({
               id="site-deploy"
               key="site-deploy"
               title="Deploy"
-              accessoryTitle="This will run the deploy script for your site"
+              accessories={[{text: "This will run the deploy script for your site"}]}
               icon={Icon.Hammer}
               actions={
                 <ActionPanel>
@@ -111,7 +110,7 @@ export const SitesSingleView = ({
             id="site-flush-fastcgi-cache"
             key="site-flush-fastcgi-cache"
             title="Flush FastCGI Cache"
-            accessoryTitle="This will flush the FastCGI cache"
+            accessories={[{text: "This will flush the FastCGI cache"}]}
             icon={Icon.ArrowClockwise}
             actions={
               <ActionPanel>
@@ -128,7 +127,7 @@ export const SitesSingleView = ({
             key="open-in-ssh"
             title={`Open SSH Connection (${site.systemUser})`}
             icon={Icon.Terminal}
-            accessoryTitle={`ssh://${site.systemUser}@${server.ipAddress}`}
+            accessories={[{text: `ssh://${site.systemUser}@${server.ipAddress}` }]}
             actions={
               <ActionPanel>
                 <Action.OpenInBrowser
@@ -143,7 +142,7 @@ export const SitesSingleView = ({
             key="open-in-sftp"
             title={`Open SFTP Connection (${site.systemUser})`}
             icon={Icon.Terminal}
-            accessoryTitle={`sftp://${site.systemUser}@${server.ipAddress}`}
+            accessories={[{text: `sftp://${site.systemUser}@${server.ipAddress}`}]}
             actions={
               <ActionPanel>
                 <Action.OpenInBrowser
@@ -160,7 +159,7 @@ export const SitesSingleView = ({
             key="open-in-ploi"
             title="Open In ploi.io"
             icon={Icon.Globe}
-            accessoryTitle="ploi.io"
+            accessories={[{text: "ploi.io"}]}
             actions={
               <ActionPanel>
                 <Action.OpenInBrowser
@@ -185,7 +184,7 @@ export const SitesSingleView = ({
                   id={key}
                   key={key}
                   title={label}
-                  accessoryTitle={value}
+                  accessories={[{text: value}]}
                   icon={Icon.Document}
                   actions={
                     <ActionPanel>