+++ extensions/ploi/src/api/Server.ts
@@ -1,14 +1,19 @@
 import { LocalStorage, showToast, Toast, popToRoot } from "@raycast/api";
 import { IServer } from "../Server";
 import { camelCase, mapKeys, sortBy } from "lodash";
-import { PLOI_API_URL } from "../config";
+import { PLOI_API_URL, PLOI_PER_PAGE } from "../config";
 import axios, { AxiosError } from "axios";
+import { PaginatedResponse } from "../common/types";
 
 export const Server = {
-  async getAll() {
-    const servers = await getServers();
+  async getAll(page: number) {
+    const response = await getServers(page);
 
-    return sortBy(servers, (s) => s.name.toLowerCase()) ?? {};
+    const sortedServers = sortBy(response?.data, (s) => s.name.toLowerCase()) ?? {};
+    return {
+      ...response,
+      data: sortedServers
+    };
   },
 
   async reboot({
@@ -77,16 +82,20 @@ export const Server = {
   },
 };
 
-const getServers = async () => {
+const getServers = async (page: number) => {
   try {
-    const response = await axios.get(`${PLOI_API_URL}/servers?per_page=50`);
+    const response = await axios.get(`${PLOI_API_URL}/servers?per_page=${PLOI_PER_PAGE}&page=${page}`);
 
     const serverData = (await response.data) as ServersResponse;
     const servers = serverData?.data ?? [];
-
+    
     // eslint-disable-next-line
     // @ts-expect-error Not sure how to convert Dictionary from lodash to IServer
-    return servers.map((s) => mapKeys(s, (_, k) => camelCase(k)) as IServer);
+    const mappedServers = servers.map((s) => mapKeys(s, (_, k) => camelCase(k)) as IServer);
+    return {
+      ...serverData,
+      data: mappedServers
+    };
   } catch (error) {
     const axiosError = (error as AxiosError).response;
 
@@ -109,6 +118,4 @@ const getServers = async () => {
   }
 };
 
-type ServersResponse = {
-  data: IServer[];
-};
+type ServersResponse = PaginatedResponse<IServer>;