+++ extensions/ploi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Ploi Changelog
 
+## [Paginated Servers] - {PR_MERGE_DATE}
+
+- You can now view Servers even if they are > 50 (Pagination for Sites will come later)
+- Updated metadata images
+- `chore` and updates
+
 ## [New features] - 2022-06-09
 
 - Updated to latest Raycast API